Refactored app into seperate componenets
