+++ src/Form/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "./Form.js",
+  "name": "Form",
+  "version": "0.0.0",
+  "private": true
+}
