+++ src/Header/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "./Header.js",
+  "name": "Header",
+  "version": "0.0.0",
+  "private": true
+}
