+++ src/Info/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "./Info.js",
+  "name": "Info",
+  "version": "0.0.0",
+  "private": true
+}
