+++ src/RepliesAndForwards/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "./RepliesAndForwards.js",
+  "name": "RepliesAndForwards",
+  "version": "0.0.0",
+  "private": true
+}
