+++ src/SignatureContainer/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "./SignatureContainer.js",
+  "name": "SignatureContainer",
+  "version": "0.0.0",
+  "private": true
+}
