+++ src/constants/index.js
@@ -1,0 +1,20 @@
+export const brandInfo = {
+  brandName: 'Readify',
+  brandLink: 'https://readify.net',
+  brandLinkName: 'readify.net',
+  brandLogo: {
+    link: 'https://readifysignatures.blob.core.windows.net/images/image001.png',
+    alt: 'Readify a Telstra Company'
+  },
+  supportEmail: 'support@readify.net',
+  supportMobile: '&nbsp;1800&nbsp;READIFY'
+};
+
+export const placeholders = {
+  name: 'Your Name',
+  title: 'Job Title',
+  email: 'your.name@readify.net',
+  mobile: '+61 111 222 333',
+  twitter: '@myTwitter',
+  qualifications: 'Jedi Master | PSM I'
+};
